TechBook/SkillsManager_Examples/requirements.txt
python-dotenv
openai
google-genai
requests
SkillsManager
BitSig
